work on gsh man page, first version of French manpage, italian templates
NOTE for italian translator

The files are located in $GSH_ROOT/i18n/source/gsh-commands/it

The files
  commands_admin.txt
  commands_dev.txt
  commands_misc.txt
  commands_standard.txt
were extracted from the $GSH_ROOT/i18n/gsh-full-help/it.txt and
shouldn't require modifications. (Note that the trailing "+" symbols are
there to indicate hard line breaks and should be kept.)

The file
  gsh_help.txt
is the header used with commands_standard.txt to generate the output of
the "gsh help" command. It was extracted from the
$GSH_ROOT/i18n/gsh-help/it.txt and shouldn't require modifications.

That leaves 2 files:

  gsh_full_help.txt
  It is used to wrap the commands_*.txt files to generate the output
  of the "gsh HELP" command.
  You can translate the section titles and the reference to the
  manpage. (If you don't feel an italian manpage is important, you can
  either remove the first few lines, or say that the manpage is in
  English.)

  man_gsh.txt
  It is used to generate the manpage (with asciidoc) with the
  commands_*.txt files.
  It contains additional comments / explanations about GameShell. If
  you do translate it, don't forget to add your name somewhere!

You can generate each of the output files with

  $ make it.1
  $ make it.help

The command

  $ make install

should install the files in the appropriates directories.

NOTE that the manpage is still a work in progress, and I'm still not
sure that's the final version. Translating it is thus not a priority at
this point.